Add alternative DQN reinforcement learning implementation (#175)
* updating gitignore for IDEA support

* implement basic DQN

* adding learningRateDecay and epsilonDecay
adding a new util-class "Window", works similar as the Java ArrayDeque
add training mode for the DQN
much improvement in performance
allow mulit-hidden layers

* fixing test method

* changing test method

* add comments to the DQN class

* removing debug output

* adding required import

* reducing complexity

* reducing complexity#2

* fixing bug in DQN.learn()

* Fix act method for high initial exploration

The sign for the epsilon action comparison was reversed. Epsilon should begin as a high number which will lead to high exploration at first. By exploring at a high rate the agent can discover new states instead of following its judgement (poor at first) about what the best state is and with a decay function over time epsilon is decreased which then leads the agent to trust its judgement more as it has more experiences.

* Change 'epsilon' properties to 'explore'

To be as beginner friendly as possible, we change 'epsilon'-based property names to 'explore'. We do this because the people aware of epsilon's meaning should be able to know that explore is an equivalent term, but the reverse may not be true

* Change 'epsilon' properties to 'explore' 

To be as beginner friendly as possible, we change 'epsilon'-based property names to 'explore'. We do this because the people aware of epsilon's meaning should be able to know that explore is an equivalent term, but the reverse may not be true

* Refactor 'DQN.act', Add description details